# thesishub-application/api-docs/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.1-SNAPSHOT",
        "title":"Thesis Hub Application"
    },
    "host":"localhost:8080",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
        "/account/activate":{
            "post":{
                "consumes":[
                    "application/x-www-form-urlencoded"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"query",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"secret",
                        "in":"query",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"password",
                        "in":"formData",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/confirmReset":{
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"query",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"secret",
                        "in":"query",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/invite":{
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "headers":{
                            "Location":{
                                "type":"string"
                            }
                        },
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    },
                    "409":{
                        "description":"Conflict",
                        "headers":{
                        }
                    },
                    "500":{
                        "description":"Internal Server Error",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/resendActivation":{
            "post":{
                "consumes":[
                    "application/x-www-form-urlencoded"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"email",
                        "in":"formData",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/resetPassword":{
            "post":{
                "consumes":[
                    "application/x-www-form-urlencoded"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"email",
                        "in":"formData",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/signUp":{
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "headers":{
                            "Location":{
                                "type":"string"
                            }
                        },
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "headers":{
                        }
                    },
                    "409":{
                        "description":"Conflict",
                        "headers":{
                        }
                    },
                    "500":{
                        "description":"Internal Server Error",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/account/{id}/password":{
            "put":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/UpdatePasswordBean"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "headers":{
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
        "Company":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "plan":{
                    "$ref":"#/definitions/CompanyPlan"
                },
                "size":{
                    "type":"string",
                    "enum":[
                        "CORPORATE",
                        "MEDIUM",
                        "SMALL",
                        "STARTUP"
                    ]
                },
                "url":{
                    "type":"string"
                }
            }
        },
        "CompanyPlan":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "maxTopics":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "Country":{
            "properties":{
                "name":{
                    "type":"string"
                },
                "tag":{
                    "type":"string"
                }
            }
        },
        "Faculty":{
            "properties":{
                "id":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                },
                "university":{
                    "$ref":"#/definitions/University"
                }
            }
        },
...
        "University":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "url":{
                    "type":"string"
                }
            }
        },
        "UpdatePasswordBean":{
            "properties":{
                "newPwd":{
                    "type":"string"
                },
                "oldPwd":{
                    "type":"string"
                }
            }
        },
        "User":{
            "properties":{
                "company":{
                    "$ref":"#/definitions/Company"
                },
                "email":{
                    "type":"string"
                },
                "faculty":{
                    "$ref":"#/definitions/Faculty"
                },
                "id":{
                    "type":"integer"
                },
                "lastLogin":{
                    "type":"object"
                },
                "name":{
                    "type":"string"
                },
                "phone":{
                    "type":"string"
                },
                "roles":{
                    "type":"string",
                    "enum":[
                        "AC_SUPERVISOR",
                        "ADMIN",
                        "COMPANY_REP",
                        "PROJECT_LEADER",
                        "STUDENT",
                        "TECH_LEADER",
                        "UNIVERSITY_AMB"
                    ]
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "username":{
                    "type":"string"
                }
            }
        }
    }
}